Implemented ChannelStreaming Consumer on the server.  Saving of channel data chunks still pending.
Related Work Items: #5060

+++ src/Witsml.Studio.Plugins.DataReplay/ViewModels/Proxies/Etp141ProxyViewModel.cs
@@ -182,7 +182,7 @@ namespace PDS.Witsml.Studio.Plugins.DataReplay.ViewModels.Proxies
 
             var value = 0d;
 
-            if (streamingInfo.StartIndex.Item is long)
+            if (streamingInfo.StartIndex.Item is double)
             {
                 value = (double)streamingInfo.StartIndex.Item 
                       + Math.Pow(10, index.Scale);
